+++ signature/src/bls_over_bn254.rs
@@ -171,11 +171,15 @@ impl AggregateableSignatureSchemes for BLSOverBN254CurveSignatureScheme {
                 msgs.len(),
             )));
         }
-        // subgroup check
-        // TODO: for BN we don't need a subgroup check
-        sig.sigma.check().map_err(|_e| {
-            SignatureError::ParameterError("signature subgroup check failed".to_string())
-        })?;
+        // both oncurve check and subgroup check for verification keys
+        vks.iter()
+            .try_for_each(|vk| vk.check().map_err(|_| SignatureError::FailedValidityCheck))?;
+        // NOTE: for BN curve, we don't need subgroup check, since co-factor is 1,
+        // thus only conducting on_curve check
+        if !sig.sigma.into_affine().is_on_curve() {
+            return Err(SignatureError::FailedOnCurveCheck);
+        }
+
         // verify
         let mut m_points: Vec<G1Prepared<_>> = msgs
             .iter()
@@ -215,6 +219,15 @@ impl AggregateableSignatureSchemes for BLSOverBN254CurveSignatureScheme {
                 "no verification key for signature verification".to_string(),
             ));
         }
+        // both oncurve check and subgroup check for verification keys
+        vks.iter()
+            .try_for_each(|vk| vk.check().map_err(|_| SignatureError::FailedValidityCheck))?;
+        // NOTE: for BN curve, we don't need subgroup check, since co-factor is 1,
+        // thus only conducting on_curve check
+        if !sig.sigma.into_affine().is_on_curve() {
+            return Err(SignatureError::FailedOnCurveCheck);
+        }
+
         let mut agg_vk = vks[0].0;
         for vk in vks.iter().skip(1) {
             agg_vk += vk.0;
@@ -510,6 +523,11 @@ impl VerKey {
         sig: &Signature,
         csid: B,
     ) -> Result<(), SignatureError> {
+        // NOTE: for BN curve, we don't need subgroup check, since co-factor is 1,
+        // thus only conducting on_curve check
+        if !sig.sigma.into_affine().is_on_curve() {
+            return Err(SignatureError::FailedOnCurveCheck);
+        }
         let msg_input = [msg, csid.as_ref()].concat();
         let group_elem = hash_to_curve::<Keccak256>(&msg_input);
         let g2 = G2Projective::generator();
@@ -530,10 +548,13 @@ mod tests {
         bls_over_bn254::{BLSOverBN254CurveSignatureScheme, KeyPair, SignKey, Signature, VerKey},
         constants::CS_ID_BLS_BN254,
         tests::{agg_sign_and_verify, failed_verification, sign_and_verify},
+        AggregateableSignatureSchemes, SignatureError,
     };
+    use ark_bn254::{Fq, G1Affine};
+    use ark_ec::AffineRepr;
     use ark_ff::vec;
     use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
-    use ark_std::{vec::Vec, UniformRand};
+    use ark_std::{rand::Rng, vec::Vec, UniformRand};
 
     #[test]
     fn test_bls_signature_internals() {
@@ -660,5 +681,69 @@ mod tests {
             assert_eq!(vk.cmp(&vk_copy), ark_std::cmp::Ordering::Equal);
             assert_eq!(vk, vk_copy);
         }
+    }
+
+    #[test]
+    fn test_malformed_signature() {
+        let mut rng = jf_utils::test_rng();
+        let keypair = KeyPair::generate(&mut rng);
+        let vk = keypair.ver_key();
+        let msg = b"hello";
+
+        let bad_sig = Signature {
+            sigma: G1Affine::new_unchecked(Fq::rand(&mut rng), Fq::rand(&mut rng)).into_group(),
+        };
+        assert_eq!(
+            vk.verify(msg.as_slice(), &bad_sig, CS_ID_BLS_BN254),
+            Result::Err(SignatureError::FailedOnCurveCheck)
+        );
+        assert_eq!(
+            BLSOverBN254CurveSignatureScheme::multi_sig_verify(
+                &(),
+                &[vk],
+                msg.as_slice(),
+                &bad_sig
+            ),
+            Result::Err(SignatureError::FailedOnCurveCheck)
+        );
+        assert_eq!(
+            BLSOverBN254CurveSignatureScheme::aggregate_verify(
+                &(),
+                &[vk],
+                &[msg.as_slice()],
+                &bad_sig
+            ),
+            Result::Err(SignatureError::FailedOnCurveCheck)
+        );
+    }
+
+    #[test]
+    fn test_malformed_vk() {
+        let mut rng = jf_utils::test_rng();
+        let keypair = KeyPair::generate(&mut rng);
+        let msg = b"hello";
+        let sig = keypair.sign(msg.as_slice(), CS_ID_BLS_BN254);
+
+        let mut bad_vk = keypair.ver_key();
+        bad_vk.0.x.c0 += Fq::rand(&mut rng);
+
+        assert_eq!(
+            BLSOverBN254CurveSignatureScheme::multi_sig_verify(
+                &(),
+                &[bad_vk],
+                msg.as_slice(),
+                &sig
+            ),
+            Result::Err(SignatureError::FailedValidityCheck),
+        );
+        assert_eq!(
+            BLSOverBN254CurveSignatureScheme::aggregate_verify(
+                &(),
+                &[bad_vk],
+                &[msg.as_slice()],
+                &sig
+            ),
+            Result::Err(SignatureError::FailedValidityCheck),
+        );
     }
 }